==> 00-Geoconvert Class Testing.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import requests
import urllib
import six
import re
import os
from geoconvert import geoconvert

src = os.path.join('data','src')

if not os.path.exists(os.path.join(src,'GeoConvertTest.csv')):
    print("Extracting for testing with GeoConvert: " + os.path.join(src,'2001','ks004.csv.gz'))
    
    test = pd.read_csv(os.path.join(src,'2001','ks004.csv.gz'), 
                        header=5, skipfooter=4, skip_blank_lines=True, engine='python', compression='gzip')#, na_values="..")

    test.rename(columns={
        'super output areas - lower layer':'lsoanm',
        'mnemonic':'lsoacd',
        'All people':'total',
    }, inplace=True)
    
    test = test.loc[:,('lsoacd','total')]

    # Save it
    test.to_csv(os.path.join(src,'GeoConvertTest.csv'), index=False)

print("Done.")

gc = geoconvert.geoconvert()
gc.auto_2001_to_2011(os.path.join(src,'GeoConvertTest.csv'))

dfc = pd.read_csv(os.path.join('GeoConvertTest-converted.csv'), index_col=False)
    
dfc.describe()
